+++ nestedtensor/csrc/py_init.cpp
@@ -2,7 +2,6 @@
 #include <jit_list_apply.h>
 #include <torch/extension.h>
 #include <unary.h>
-#include <seg_layers.h>
 #include <utils/nested_node_functions.h>
 #include <utils/python_nested_node.h>
 #include <python_functions.h>
@@ -88,6 +87,8 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
       .def("__repr__", &THPNestedTensor::str);
 
   add_unary_functions(m, c);
+  add_functions(m, c);
+
   add_relu(m, c, "relu");
   add_dropout(m, c, "dropout");
   add_conv2d(m, c, "conv2d");
@@ -95,8 +96,6 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
   add_batch_norm(m, c, "batch_norm");
   add_cross_entropy(m, c, "cross_entropy");
   add_interpolate(m, c, "interpolate");
-  add_interpolate_single_size(m, c, "interpolate");
-  add_functions(m, c);
 
   // NOTE: This is a private function until it is feature complete
   m.def("_jit_tensorwise", &torch::nested_tensor::jit_tensorwise);
