+++ nestedtensor/version.py
@@ -1,5 +1,5 @@
-__version__ = '0.0.1.dev20204616+8ffeb6a'
-git_version = '8ffeb6ad6bbd4e5f5b8a0d4e997156b9eb98132f'
+__version__ = '0.0.1.dev20204620+fb11447'
+git_version = 'fb1144700d8a19532837f467b8d66c6fbf21ded2'
 from nestedtensor import _C
 if hasattr(_C, 'CUDA_VERSION'):
     cuda = _C.CUDA_VERSION
